Se agrefgaron botones a autores y la funcion de eliminar autor

+++ templates_c/2fb0e742df7b93057a56cee924f2cc42b7f2783b_0.file.libroTabla.tpl.php
@@ -1,18 +1,18 @@
 <?php
-/* Smarty version 4.5.2, created on 2024-06-04 17:24:26
+/* Smarty version 4.5.2, created on 2024-06-05 03:35:17
   from 'C:\xampp\htdocs\Proyectos\ProyectosWeb2\literatura\templates\libroTabla.tpl' */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '4.5.2',
-  'unifunc' => 'content_665f31aaac6574_97949350',
+  'unifunc' => 'content_665fc0d522a243_42598837',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
     '2fb0e742df7b93057a56cee924f2cc42b7f2783b' => 
     array (
       0 => 'C:\\xampp\\htdocs\\Proyectos\\ProyectosWeb2\\literatura\\templates\\libroTabla.tpl',
-      1 => 1717474048,
+      1 => 1717551195,
       2 => 'file',
     ),
   ),
@@ -22,7 +22,7 @@ if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
     'file:finHtml.tpl' => 1,
   ),
 ),false)) {
-function content_665f31aaac6574_97949350 (Smarty_Internal_Template $_smarty_tpl) {
+function content_665fc0d522a243_42598837 (Smarty_Internal_Template $_smarty_tpl) {
 $_smarty_tpl->_subTemplateRender('file:formAddLibro.tpl', $_smarty_tpl->cache_id, $_smarty_tpl->compile_id, 0, $_smarty_tpl->cache_lifetime, array(), 0, false);
 ?>
 <div class="container col-12">
@@ -35,7 +35,7 @@ $_smarty_tpl->_subTemplateRender('file:formAddLibro.tpl', $_smarty_tpl->cache_id
           <th scope="col">Precio</th>
           <th scope="col">Cantidad de paginas</th>
           <th scope="col">Autor</th>
-          <th scope="col">Acciones</th>
+  
           
        
         </tr>
@@ -63,12 +63,13 @@ $_smarty_tpl->tpl_vars['l']->do_else = false;
             <td><?php echo $_smarty_tpl->tpl_vars['l']->value->id_autor;?>
 </td>
             <td>
+          <?php if ($_smarty_tpl->tpl_vars['usuario']->value == 'administrador') {?>
               <a href='verunAutor/<?php echo $_smarty_tpl->tpl_vars['l']->value->id_autor;?>
 ' action='autores' class='btn btn-success'>Ver autor</a>  
               <a href='delete/<?php echo $_smarty_tpl->tpl_vars['l']->value->id_libro;?>
 ' class='btn btn-danger'>Eliminar</a>
               </td>
-           </tr>
+                         <?php }?>
         <?php
 }
 $_smarty_tpl->smarty->ext->_foreach->restore($_smarty_tpl, 1);?>
